requirements.txt
appdirs==1.4.3
boto==2.48.0
botocore==1.7.47
certifi==2017.7.27.1
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.1
django-compat==1.0.15
django-crispy-forms==1.7.2
django-filter==2.1.0
django-hosts==3.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-storages-redux==1.3.3
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
docutils==0.14
flake8==3.5.0
forex-python==1.5
future==0.17.1
geographiclib==1.49
gunicorn==19.7.1
html5lib==1.0.1
humanize==0.5.1
idna==2.6
jmespath==0.9.3
lxml==4.3.4
Markdown==3.0.1
markdown2==2.3.4
mccabe==0.6.1
numpy==1.16.2
pandas==0.24.2
psycopg2-binary==2.7.7
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.6.0
pytz==2017.2
ratelim==0.1.6
requests==2.18.4
simplejson==3.10.0
six==1.10.0
urllib3==1.22
webencodings==0.5.1
whitenoise==4.1.2
xmltodict==0.11.0
